requirements.txt
streamlit
tensorflow
Pillow
numpy
opencv-python
streamlit-drawable-canvas